[重构] notebook逻辑变更 由失败触发变更为成功触发
